+++ src/library/DIPS.Mobile.UI/Effects/Layout/iOS/LayoutPlatformEffect.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace DIPS.Mobile.UI.Effects.Layout;
 
 public partial class LayoutPlatformEffect
@@ -18,8 +20,17 @@ public partial class LayoutPlatformEffect
     {
         if(Control is null)
             return;
-        
-        Control.ClipsToBounds = m_originalClipToBound;
-        Control.Layer.CornerRadius = m_prevCornerRadius;
+
+        try
+        {
+            Control.ClipsToBounds = m_originalClipToBound;
+            Control.Layer.CornerRadius = m_prevCornerRadius;
+        }
+        catch
+        {
+            // I believe this can happen if Layer is null
+            // We can safely swallow this, as the Control is no longer visible anyway when Layer is null
+            // The idea here was to reset the CornerRadius if consumer removes the effect while the Control is still visible
+        }
     }
 }